+++ jshERP-boot/src/main/java/com/jsh/erp/controller/InOutItemController.java
@@ -23,6 +23,8 @@ import java.util.Map;
 
 import static com.jsh.erp.utils.ResponseJsonUtil.returnJson;
 import static com.jsh.erp.utils.ResponseJsonUtil.returnStr;
+import com.jsh.erp.service.UserService;
+import static com.jsh.erp.utils.ResponseJsonUtil.returnForbidden;
 
 /**
  * @author jishenghua  jshERP 2018年12月25日14:38:08
@@ -32,6 +34,9 @@ import static com.jsh.erp.utils.ResponseJsonUtil.returnStr;
 @Api(tags = {"收支项目"})
 public class InOutItemController extends BaseController {
     private Logger logger = LoggerFactory.getLogger(InOutItemController.class);
+
+        @Resource
+    private UserService userService;
 
     @Resource
     private InOutItemService inOutItemService;
@@ -64,6 +69,7 @@ public class InOutItemController extends BaseController {
     @PostMapping(value = "/add")
     @ApiOperation(value = "新增")
     public String addResource(@RequestBody JSONObject obj, HttpServletRequest request)throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         int insert = inOutItemService.insertInOutItem(obj, request);
         return returnStr(objectMap, insert);
@@ -72,6 +78,7 @@ public class InOutItemController extends BaseController {
     @PutMapping(value = "/update")
     @ApiOperation(value = "修改")
     public String updateResource(@RequestBody JSONObject obj, HttpServletRequest request)throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         int update = inOutItemService.updateInOutItem(obj, request);
         return returnStr(objectMap, update);
@@ -80,6 +87,7 @@ public class InOutItemController extends BaseController {
     @DeleteMapping(value = "/delete")
     @ApiOperation(value = "删除")
     public String deleteResource(@RequestParam("id") Long id, HttpServletRequest request)throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         int delete = inOutItemService.deleteInOutItem(id, request);
         return returnStr(objectMap, delete);
@@ -88,6 +96,7 @@ public class InOutItemController extends BaseController {
     @DeleteMapping(value = "/deleteBatch")
     @ApiOperation(value = "批量删除")
     public String batchDeleteResource(@RequestParam("ids") String ids, HttpServletRequest request)throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         int delete = inOutItemService.batchDeleteInOutItem(ids, request);
         return returnStr(objectMap, delete);
@@ -148,6 +157,7 @@ public class InOutItemController extends BaseController {
     @ApiOperation(value = "批量设置状态")
     public String batchSetStatus(@RequestBody JSONObject jsonObject,
                                  HttpServletRequest request)throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Boolean status = jsonObject.getBoolean("status");
         String ids = jsonObject.getString("ids");
         Map<String, Object> objectMap = new HashMap<>();
