+++ jshERP-boot/src/main/java/com/jsh/erp/controller/SysDictDataController.java
@@ -21,6 +21,7 @@ import java.util.List;
 import java.util.Map;
 
 import static com.jsh.erp.utils.ResponseJsonUtil.returnStr;
+import static com.jsh.erp.utils.ResponseJsonUtil.returnForbidden;
 
 /**
  * 数据字典信息
@@ -80,6 +81,7 @@ public class SysDictDataController extends BaseController {
     @ApiOperation(value = "新增字典数据")
     @PostMapping(value = "/add")
     public String add(@Validated @RequestBody SysDictData dict) throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         dict.setCreateBy(userService.getCurrentUser().getLoginName());
         return returnStr(objectMap, dictDataService.insertDictData(dict));
@@ -91,6 +93,7 @@ public class SysDictDataController extends BaseController {
     @ApiOperation(value = "修改保存字典数据")
     @PutMapping(value = "/update")
     public String edit(@Validated @RequestBody SysDictData dict) throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         dict.setUpdateBy(userService.getCurrentUser().getLoginName());
         return returnStr(objectMap, dictDataService.updateDictData(dict));
@@ -101,7 +104,8 @@ public class SysDictDataController extends BaseController {
      */
     @DeleteMapping("/{dictCodes}")
     @ApiOperation(value = "删除字典数据")
-    public AjaxResult remove(@PathVariable Long[] dictCodes) {
+    public AjaxResult remove(@PathVariable Long[] dictCodes) throws Exception {
+        if (!userService.isCurrentUserAdmin()) { return error("无权限"); }
         dictDataService.deleteDictDataByIds(dictCodes);
         return success();
     }
@@ -109,6 +113,7 @@ public class SysDictDataController extends BaseController {
     @DeleteMapping(value = "/delete")
     @ApiOperation(value = "删除")
     public String deleteResource(@RequestParam("id") Long id, HttpServletRequest request)throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         int delete = dictDataService.deleteDictData(id, request);
         return returnStr(objectMap, delete);
@@ -117,6 +122,7 @@ public class SysDictDataController extends BaseController {
     @DeleteMapping(value = "/deleteBatch")
     @ApiOperation(value = "批量删除")
     public String batchDeleteResource(@RequestParam("ids") String ids, HttpServletRequest request)throws Exception {
+        if (!userService.isCurrentUserAdmin()) return returnForbidden();
         Map<String, Object> objectMap = new HashMap<>();
         int delete = dictDataService.batchDeleteDictData(ids, request);
         return returnStr(objectMap, delete);
